BUG: Targets shape correction in hybrid dropout
hybrid_dropout.py has been corrected in order to detect the dimensions of the targets. When the target dimension is not equal to 1 the dropout will raise an error especifiying the expected shape. The expected shape is not [batch_size, num_classes], instead it will be just [batch_size]

Also test_hybrid_dropout.py has been improved to check this target shape